refactor(apps): createApps is an assembler, not a 2,600-line function
Every private helper and every door `createApps` returned moves into a
module beside its contract, following the shape #963 gave the namespace
surfaces: a module taking a `Pick` of the one shared closure type and
returning its slice.

`createApps` is 17 lines. The longest function anywhere in the touched
set is 113. Nothing about the package's exports changed and no test
file changed.
